Remove allocator-uses-construction support from core::any
Reasons for this are due to the lack of allocator-uses-construction
support in libstdc++ as of g++ 4.8.0, which is currently the minimum
supported version for MNMLSTC Core for g++.

While allocator-uses-construction is extremely valuable and part of the
proposal, lack of current library support means it will be difficult to
test and insure that the logic is sane. It will be added in into a
future version of core::any, when a minimum version bump occurs.