Add 'unpack'-able invoke, which unpacks a tuple with perfect forwarding.
Clean up the type traits a bit.
Fix a typo in functional for an invoke signature.

The function signature for the unpack invoke may change. I don't know
how I feel about it going 'unpack', 'call', 'args'. Ideally it should be
'call', 'unpack', 'args'. This would then let a invoke of this invoke
take place, and I like this flexibility.