Change semantics for runtime unpacking.
Calling core::invoke with the 'runpack' sentinel will use
function_traits to figure out the arity of the functor and then generate
the necessary index_sequence based off of this information. This has
resulted in the removal of runtime unpacking an entire container that
would take the first element in the container, then unpack the rest of the
container into the first element. It is considered ok for this obscure
use to be lost