Fix issue with invoke and constexpr
I should pay more attention to what can and cannot be constexpr in
C++11. std::forward<T> is *not* constexpr and therefore my logic breaks
down in these cases.

I can always add the features for unpacking and runpacking to be
constexpr in c++14 which will be the v2 namespace for this library.