Close #1 (with one caveat, see below)
Finish implementing behavior in the optional component to conform to N3793
Update optional component documentation

Due to a bug in Clang 503.0.40 (the version of clang currently provided
with XCode 5.1.1 on OS X), make_optional has not been marked as
constexpr. I do not expect this to change for some time. While it may
work in the latest clang built from source, many people will simply use
the XCode supplied version of clang, and they will get bizarre errors
when trying to use a constexpr make_optional. Such is life. A note has
been made in the documentation. If a minimum required version bump of
MNMLSTC Core occurs, then this may be revisited.