Close #12
core::invoke is now usable with constexpr values. PLEASE NOTE: GCC 4.8.1
currently has a bug with function pointers losing their constexpr-ness
when passed to a constexpr function. This is completely out of my
control. However, it also severely limits the usage of core::invoke in
some situations, as frustrating as it may be.