Implement some changes for string_view (Issue #33)
  The substr functions and a few others are not yet implemented
correctly. These will follow soon. Please be aware, even though
basic_string_view is not DLL exportable (as defined by core), the ABI
has changed for the find functions (due to the signature change).
Behavior should stay the same when compared to previous versions.

Add 'meta' module

  This is where all future meta template functions will be placed. The
previously (poorly) named 'all_traits', 'any_traits', and 'none_traits',
are no deprecated (and simply aliases to the meta function versions).
Of course I discovered that you can't deprecate template aliases with
attributes. Wamp wamp. :/

Add 'void_t'-like type (this is named deduce_t because I personally
think that's a better name)

Add a whole mess of type traits, mostly operator related.
  These are in the 'trait' namespace within core. They may get moved to
their own header, because of just how many there are, and I don't see
them being used as often as the rest of the type_traits header. On the
bright side, core::trait implements checking for the subscript operator,
the comma operator, and the arrow operator. Additionally, you can also
check for the addressof operator as a member.

Fix a compiler error with GCC in core::any that seems to have slipped
past for a while.

This is a fairly decently sized change, and is frankly the cause for the
1.2 release being delayed.