Fix CMakeLists.txt
To link against libc++ the library has to appear in the link options - default is libstdc++ which leads to undefined references.

Further the check for c++11 has to be done every time - without that check the USE_STD_CXX11 variable is not set and the include header are not found.

This was tested with clang 3.6.2 and cmake 3.2.2 on Ubuntu 14.04