Clean up Travis test coverage
This patch revamps ci builds by adding support for a gcc (4.8, 4.9, 5, and 6), clang (3.4, 3.5, 3.6, 3.8, and 3.9), and Xcode (6.4, 7.3, and 8). It strips out many unnecessary components in the setup/build scripts, and adds code coverage with codecov.io.

Additionally, it removes a check from the any tests that is effectively undefined behavior and definitely fails on many release builds.