Fix murmur implementation for 32 bit architectures
The current usage of the murmur<8> struct is valid, but only because it is used for 64 bit architectures. On 32 bit architectures, sizeof(size_t) is less than the shift amount and results in a great number of warnings when compiled on those platforms.

In order to allow the murmur<8> implementation to function correctly on 32 bit architectures, the hash size has been set to an architecture-independent size of 64 bits as this is the correct size for that implementation.