Fix core::variant when using libc++ and std=c++1z
As C++17 has introduced std::invoke, Xcode 8 on macOS, possibly erroneously, states that there is an ambiguous function call in the implementation of core::variant when compiling with -std=c++1z and -stdlib=libc++. To fix the issue, the fully qualified name, ::core::v2::invoke, has been used.

To ensure that similar issues do not occur in the future, two additional CI jobs have been created on Travis to build the project in these conditions.